safari-600.7-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.2_White_Space/S7.2_A5_T4-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Invalid unicode escape in identifier: '\u0020'
S7.2_A5_T4

PASS Expected parsing failure

TEST COMPLETE
